Move RssFeed in rss module
